Fibonacci Coding
The Fibonacci code word for a particular integer is exactly the
integer’s Zeckendorf representation with the order of its digits
reversed and an additional “1” appended to the end. The extra 1 is
appended to indicate end of code The representation uses Fibonacci
numbers starting from 1 (2’nd Fibonacci Number). So the Fibonacci
Numbers used are 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 141, …….

Given a number n, print its Fibonacci code.